+++ examples/game3d-experiments/src/index.ts
@@ -1,5 +1,6 @@
 import { Entity } from "@bigby/core";
 import { Game3D } from "@bigby/3d";
+import { Ticker } from "@bigby/behaviors";
 
 import "./index.css";
 
@@ -7,6 +8,7 @@ const canvas = document.getElementById("game") as HTMLCanvasElement;
 console.log(canvas);
 
 const game = new Entity();
+game.addBehavior(Ticker);
 game.addBehavior(Game3D);
 game.getBehavior(Game3D).canvas = canvas;